refactor: move component extension
