+++ AGENTS.md
@@ -8,13 +8,22 @@ Markpage helps to render Markdown files on html pages with any framework.
 
 You point Markpage at a directory with markdown files and get navigation structure and content that you can use to render in your app.
 
+## Monorepo structure
+
+This repository contains multiple packages:
+
+- `packages/markpage` – core builder, renderer and types
+- `packages/markpage-svelte` – Svelte integration (components in markdown)
+- `packages/website` – documentation site built with SvelteKit and Markpage
+- `packages/tests` – test suite covering builder/renderer and Svelte integration
+
 ## What it does
 
 Point Markpage at a directory with markdown files and `.index.json` files, and get:
 - **Organized navigation structure** for your content
 - **Multiple output formats** (app bundles, website navigation, static HTML)
 - **Framework-agnostic** utilities that work with React, Vue, Svelte, Angular, or vanilla JavaScript
-- **Component system** for embedding interactive components in markdown (Svelte support included)
+- **Optional component system** for embedding interactive components in markdown (via separate framework packages; Svelte support available)
 
 ## Quick Start
 
@@ -31,9 +40,23 @@ await buildPages('./my-content', {
 });
 ```
 
-## Component System
+### Scripts (root)
 
-Markpage supports embedding interactive components directly in markdown files, similar to MDX but simpler:
+Useful root scripts for local development:
+
+- `npm test` – builds local packages and runs the test suite (`packages/tests`)
+- `npm run build` – builds local packages then the documentation website
+- `npm run dev:website` – starts the SvelteKit docs website (auto-builds content from `./docs`)
+- `npm run build:website` – production build of the website
+
+Inside `packages/website`:
+
+- `npm run check` – Svelte type checks (`svelte-check`) for the website
+- `npm run preview` – preview production build locally
+
+## Component System (via @markpage/svelte)
+
+Component embedding is provided by the separate `@markpage/svelte` package (React version is not available yet — contributions welcome). It lets you embed interactive Svelte components directly in markdown files, similar to MDX but simpler:
 
 ```markdown
 # My Documentation
@@ -44,27 +67,64 @@ Here's a regular paragraph.
 <TestButton /> <!-- Uses default props -->
 ```
 
-Components are registered and used with the `MarkdownRenderer` component:
+Components are registered and rendered with the `Markdown` component:
 
 ```svelte
 <script>
-  import { MarkdownRenderer } from '@markpage/svelte';
+  import { Markdown } from '@markpage/svelte';
   import TestButton from './TestButton.svelte';
-  
+
   const components = new Map([
     ['TestButton', TestButton]
   ]);
 </script>
 
-<MarkdownRenderer 
-  content={markdownContent} 
-  components={components} 
+<Markdown 
+  source={markdownContent}
+  components={components}
 />
 ```
 
+### Framework Integrations
+
+- Svelte: `@markpage/svelte` is available on npm.
+
+```bash
+npm install @markpage/svelte
+```
+
+- React: not available yet — feel free to contribute an official `@markpage/react` integration (PRs welcome).
+
 ## Getting Started
 
-For detailed step-by-step instructions, see the [Getting Started Guide](https://github.com/mitkury/markpage/blob/main/docs/getting-started.md).
+For detailed step-by-step instructions, see the [Getting Started Guide](docs/getting-started.md).
+
+## Developing locally
+
+1) Install dependencies
+
+```bash
+npm ci
+```
+
+2) Run tests (kept green)
+
+```bash
+npm test
+```
+
+3) Run the website (auto-builds docs from `./docs`)
+
+```bash
+npm run dev:website
+```
+
+4) Build and preview the website
+
+```bash
+npm run build:website
+npm --workspace=@markpage/website run preview
+```
 
 ## Examples
 
@@ -161,20 +221,80 @@ const navigation = new NavigationTree(navigationData);
 const content = await loadContent('getting-started.md', contentBundle);
 ```
 
-## Component System
+## Component System (via @markpage/svelte)
 
-Markpage supports components in markdown files! Here are some examples:
+Markpage supports components in markdown files via the `@markpage/svelte` package. Render markdown with the `Markdown` component and provide a `components` map for custom tags or token overrides.
+
+### TestButton Component
 
 <TestButton variant="primary" text="Primary Button" />
 <TestButton variant="success" text="Success Button" />
 <TestButton variant="warning" text="Warning Button" />
 <TestButton variant="danger" text="Danger Button" />
 
-Components without props work too:
-<TestButton />
-<TestButton />
+### Button Component
 
-Components are registered upfront and can receive props like `variant` and `text`.
+<Button variant="primary">Primary Button</Button>
+<Button variant="secondary">Secondary Button</Button>
+<Button variant="danger">Danger Button</Button>
+
+### Alert Component
+
+<Alert variant="info">
+  This is an informational alert with **markdown** content inside.
+</Alert>
+
+<Alert variant="success">
+  This is a success message! Components work perfectly.
+</Alert>
+
+### Card Component
+
+<Card title="Component Features" subtitle="What you can do">
+  - Use any markdown syntax inside components
+  - Pass props to customize appearance
+  - Nest components within each other
+  - Maintain full markdown formatting
+</Card>
+
+### Unknown Component Fallback
+
+When a component isn't registered, it's rendered as plain text instead of showing an error:
+
+<UnknownComponent variant="demo" title="This component doesn't exist">
+  This content will be displayed as plain text
+</UnknownComponent>
+
+<AnotherMissingComponent size="large" />
+
+Components are registered upfront and can receive props like `variant`, `size`, `title`, etc.
+
+### Usage in Svelte
+
+```svelte
+<script lang="ts">
+  import { Markdown } from '@markpage/svelte';
+  import Button from './Button.svelte';
+  import Alert from './Alert.svelte';
+
+  const components = new Map([
+    ['Button', Button],
+    ['Alert', Alert]
+  ]);
+
+  const markdown = `
+  ## Examples
+
+  <Button variant="primary">Primary Button</Button>
+
+  <Alert variant="info">
+    This is an informational alert with **markdown** content inside.
+  </Alert>
+  `;
+</script>
+
+<Markdown source={markdown} components={components} />
+```
 
 ## CLI Usage
 
@@ -321,9 +441,10 @@ MIT License - see [LICENSE](https://github.com/mitkury/markpage/blob/main/LICENS
     "packages/*"
   ],
   "scripts": {
-    "build": "npm run copy-readme && npm run build --workspace=markpage",
+    "build": "npm run build:pkgs && npm run build:website",
+    "build:pkgs": "npm run copy-readme && npm run build --workspace=markpage && npm run build --workspace=@markpage/svelte",
     "dev": "npm run dev --workspace=markpage",
-    "test": "npm run test --workspace=@markpage/tests",
+    "test": "npm run build:pkgs && npm run test --workspace=@markpage/tests",
     "test:watch": "npm run test:watch --workspace=@markpage/tests",
     "test:coverage": "npm run test:coverage --workspace=@markpage/tests",
     "build:website": "npm run build --workspace=@markpage/website",
