+++ packages/markpage/src/renderer/index.ts
@@ -2,4 +2,4 @@ export { NavigationTree, createNavigationTree } from './navigation.js';
 export { extractHeadings, extractTableOfContents, addTableOfContents } from './content.js';
 export { builtinTokenNames } from './types.js';
 export type { BuiltinTokenName, ComponentTokenName } from './types.js';
-export { componentExtension } from './extensions/component.js';
+export { componentExtension } from '../extensions/component.js';
